prioQ: fix issue with concurrent DEQ and ENQ
introduce one cycle delay at insert_req to avoid concurrent
deq and enq access to shared register state. This is a less
optimal way to solving the problem, a better way would be
using CReg (concurrent register) to forward write data to
read port in the same cycle.

Current latency is 3 cycles, can be reduced to 2 if use CReg.